% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcC4Fraction.R
\name{calcC4Fraction}
\alias{calcC4Fraction}
\title{Calculate C4 Ratio}
\usage{
calcC4Fraction(C4.mask, GS.mask, veg.index = NULL, filename = "", ...)
}
\arguments{
\item{...}{}
}
\description{
Calculate C4 Ratio
}
