% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDel13C.R
\name{calcDel13C}
\alias{calcDel13C}
\title{Calculate d13C isoscape}
\usage{
calcDel13C(pft.cover, d13C, scale = 100, filename = "", ...)
}
\arguments{
\item{pft.cover}{RasterBrick. Percent plant functional type cover; each layer corresponds to different d13C endmember value.}

\item{d13C}{Numeric vector. Endmember values that correspond to each pft.cover layer.}

\item{scale}{Integer. Scale factor (maximum cover value). Default is 100.}

\item{filename}{Character. Optional output root filename.}

\item{...}{}
}
\value{
Raster Layer. Mean vegetation d13C.
}
\description{
Calculate d13C isoscape
}
\examples{


}
