% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPFTCover.R
\name{calcPFTCover}
\alias{calcPFTCover}
\title{Calculate Plant Functional Type Layers}
\usage{
calcPFTCover(C4.ratio, veg.layers = NULL, scale = 100, C4.flag = NULL,
  herb.flag = NULL, filename = "", ...)
}
\arguments{
\item{C4.ratio}{Raster layer. Grid cells correcpond to the C4 ratio of herbaceous cover}

\item{veg.layers}{Raster* object. Each layer corresponds to non-herbaceous cover or agriculatural crop layers}

\item{scale}{Numeric. Scale factor that represents maximum percent cover}

\item{C4.flag}{Numeric. Vector corresponding to nlayers of veg.layers object. 1 = C4 vegetation layer, 0 = C3 vegetation layer.}

\item{herb.flag}{Numeric. Vector corresponding to nlayers of veg.layers object. 1 = herbaceous vegetation layer, 0 = woody vegetation layer.}

\item{filename}{Character. Output root filename.}

\item{...}{}
}
\value{
Raster brick. Layers correspond to percent vegetation cover. Layer 1 - C4 herbaceous, Layer 2 - C3 herbaceous, Layer 3 ... - other vegetation layers provided by user.
}
\description{
Calculate Plant Functional Type Layers
}
\examples{

}
