% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMeanPre-data.R
\docType{data}
\name{COMeanPre}
\alias{COMeanPre}
\title{Mean Monthly Precipitation Grid for Colorado}
\format{An object of class \code{"RasterBrick"}, layers correspond to months;
  see \code{\link[raster]{Raster-class}}.}
\source{
\emph{CRU CL v.2.0}. Climatic Research Unit, University of East
  Anglia: \url{http://www.cru.uea.ac.uk/data/}
}
\usage{
data(COMeanPre)
}
\description{
Gridded climatology of 1961-1990 monthly mean precipitation (in mm) at 10'
  spatial resolution. Climatic Research Unit, High-Resolution Gridded
  Dataset.
}
\examples{
# Read in the data
data(COMeanPre)
\donttest{
# Plot all data layers
plot(COMeanPre)}
\donttest{
# Plot layer 3 only
plot(COMeanPre[[3]])}

}
\references{
New M, \emph{et al.} 2002. A high-resolution data set of surface
  climate over global land areas. \emph{Climate Research} 21:1-25.
}
\keyword{datasets}
