% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMonthlyNDVI-data.R
\docType{data}
\name{COMonthlyNDVI}
\alias{COMonthlyNDVI}
\title{Monthly MODIS NDVI for Colorado}
\format{An object of class \code{"RasterBrick"}, layers correspond to months;
  see \code{\link[raster]{Raster-class}}.}
\source{
\emph{MOD13A3: MODIS Vegetation Indices Monthly L3 Global 1km}. USGS
  Land Processes Distributed Active Archive Center:
  \url{https://lpdaac.usgs.gov}
}
\usage{
data(COMonthlyNDVI)
}
\description{
Monthly MODIS Normalized Difference Vegetation Index (NDVI) for Colorado in
  2001 aggregated to 10' spatial resolution.
}
\examples{
# Read in the data
data(COMonthlyNDVI)
\donttest{
# Plot all data layers
plot(COMonthlyNDVI)}
\donttest{
# Plot layer 3 only
plot(COMonthlyNDVI[[3]])}

}
\references{
Solano R, \emph{et al.} 2010. MODIS vegetation index user's guide
  (MOD13 Series). Version 2.0 (Collection 5). Vegetation Index and Phenology
  Lab, The University of Arizona, Tucson, Arizona, USA.
  \url{http://vip.arizona.edu}
}
\keyword{datasets}
