% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_months.R
\name{count_months}
\alias{count_months}
\title{Count Number of Months in Climate Mask Layers}
\usage{
count_months(climate.mask, filename = "", ...)
}
\arguments{
\item{climate.mask}{Raster* object. Mask of grid cells that meet climate
parameters, may be result of \code{ intersect_masks} or \code{mask_grids}.
Object should be  multi-layer (i.e., multi-date).}

\item{filename}{Character. Optional output root filename passed to
\code{writeRaster}, default output file type is GeoTiff. If not specified,
output is written to a temporary file.}

\item{...}{Other arguments passed to \code{writeRaster}.}
}
\value{
RasterLayer. Values correspond to number of months (or other temporal
  unit) that grid cell meets minimum climate parameters.
}
\description{
Generates raster layer of counts corresponding to number of times (usually
  months) each grid cell satisfies climate criteria (e.g., growing-season or
  C4 climate).
}
\examples{
\donttest{# Count number of months each grid cell meets C4 climate criteria

# For this example, annual stack of monthly C4 climate masks were generated
#   using mask_climate()
C4_month_total <- count_months(C4_mask)

# Plot monthly totals
plot(C4_month_total)
}
}
\seealso{
\link[grassmapr]{mask_climate}, \link[grassmapr]{mask_grids},
  \link[grassmapr]{intersect_masks}, \link{sum}.
}
