% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countC4Months.R
\name{countC4Months}
\alias{countC4Months}
\title{Count Number of C4 Months in Climate Mask Layers}
\usage{
countC4Months(C4.mask, filename = "", ...)
}
\arguments{
\item{C4.mask}{Raster* object. Mask of grid cells that meet C4 climate requirements. Object should be  multi-layer (i.e., multi-date).}

\item{filename}{Character. Output base filename (required for large files).}

\item{...}{}
}
\value{
RasterLayer. Pixel values correspond to number of months C4 climate requirements are met.
}
\description{
Count Number of C4 Months in Climate Mask Layers
}
\examples{

}
