% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropC4NA-data.R
\docType{data}
\name{cropC4NA}
\alias{cropC4NA}
\title{Percent C4 Crop Cover for North America}
\format{An object of class \code{"RasterLayer"}; see
  \code{\link[raster]{Raster-class}}.}
\source{
\emph{Cropland and Pasture Area in 2000} and \emph{Harvest Area and
  Yield for 175 Crops}. EarthStat.org: Global Landscapes Initiative,
  University of Minnesota Institute on the Environment and Ramankutty Lab,
  University of British Columbia, Vancouver: \url{http://www.earthstat.org/}
}
\usage{
data(cropC4NA)
}
\description{
Gridded percent C4 crop cover layer for North America in the year 2000 at 5'
  spatial resolution. Integrates geographic distribution of global
  agricultural lands (Ramankutty \emph{et al.} 2008) with distribution of
  crop types (Monfreda \emph{et al.} 2008), following rules described in
  Powell \emph{et al.} 2012.
}
\examples{
# Read in the data
data(cropC4NA)
\donttest{
# Plot data layer
plot(cropC4NA)}

}
\references{
Monfreda C, Ramankutty N, and Foley JA. 2008. Farming the planet:
  2. Geographic distribution of crop areas, yields, physiological types, and
  net primary production in the year 2000. \emph{Global Biogeochemical
  Cycles} 22:GB1022.

Powell RL, Yoo EH, and Still CJ. 2012. Vegetation and soil carbon-13
  isoscapes for South America: integrating remote sensing and ecosystem
  isotope measurements. \emph{Ecosphere} 3:109.

Ramankutty N., Evan AT, Monfreda C., and Foley JA. 2008. Farming the planet:
  1. Geographic distribution of global agricultural lands in the year 2000.
  \emph{Global Biogeochemical Cycles} 22:GB1003.
}
\keyword{datasets}
