% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossover.R
\name{crossover}
\alias{crossover}
\title{Mask climate Raster values}
\usage{
crossover(climate.stack, threshold, filename = "", ...)
}
\arguments{
\item{climate.stack}{Raster* object. Monthly values for a single climate
variable.}

\item{threshold}{Numeric. Threshold value for a single climate variable.}

\item{filename}{Character. Output filename (optional)}

\item{...}{}
}
\value{
RasterBrick. One output mask layer for each input data layer.
}
\description{
This function builds monthly climate masks based on a user-specified
threshold for a single climate parameter (e.g., precipitation, temperature,
 etc.).
}
\examples{

}
