% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_masks.R
\name{intersect_masks}
\alias{intersect_masks}
\title{Intersect Two Masks}
\usage{
intersect_masks(temp.mask, precip.mask, filename = "", ...)
}
\arguments{
\item{temp.mask}{Raster* object. Mask of grid cells that meet minimum
temperature threshold, may be result of  \code{mask_grids}. Object may
be single or multi-layer.}

\item{precip.mask}{Raster* object. Mask of grid cells that meet minimum
precipitation threshold, may be result of  \code{mask_grids}. Object
may be single or multi-layer.}

\item{filename}{Character. Optional output root filename passed to
\code{writeRaster}, default output file type is GeoTiff. If not specified,
output is written to a temporary file.}

\item{...}{Other arguments passed to \code{writeRaster}.}
}
\value{
RasterBrick. One output mask layer for each input layer. If filename
  is specified, output file format is GeoTiff with unsigned integer data
  type.
}
\description{
Intersects two masks to identify cells that meet both criteria for
  corresponding layers. Grid cells that satisfy both criteria are
  reclassified as 1, grid cells that fail at least one criteria are
  reclassified as 0.
}
\examples{
\donttest{# Generate masks of grid cells that satisfy C4-climate criteria
C4_mask <- intersect_masks(C4_temp_mask, precip_mask)

# Generate masks of grid cells that satisfy growing-season climate criteria
GS_mask <- intersect_masks(GS_temp_mask, precip_mask)

#Plot monthly growing-season masks
plot(GS_mask)
}
}
\seealso{
\link[grassmapr]{mask_grids}, \link[raster]{overlay}.
}
