% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred13Cisoscape.R
\name{isoscape}
\alias{isoscape}
\title{Wrapper Function to Calculate d13C isoscape}
\usage{
isoscape(temp.stack, ppt.stack, c4.threshold = 22, gs.threshold = 5,
  ppt.threshold = 25, woody = NULL, c3.crop = NULL, c4.crop = NULL,
  veg.scale = 100, veg.index = NULL, c3.endmember = -26.7,
  c4.endmember = -12.5, woody.endmember = -27.2, c3.sd = 2.3,
  c4.sd = 1.1, woody.sd = 2.5, filename = "", ...)
}
\arguments{
\item{temp.stack}{Raster* object. Each layer corresponds to a
different temporal window (e.g., month) for a single temperature variable.
Object may be single or multi-layer.}

\item{ppt.stack}{Raster* object. Each layer corresponds to a
different temporal window (e.g., month) for a single temperature variable.
Object may be single or multi-layer.}

\item{c4.threshold}{Numeric. Threshold value (lower-bound) for C4 dominance.}

\item{gs.threshold}{Numeric. Threshold value (lower-bound) of temperature
for active growing season.}

\item{ppt.threshold}{Numeric. Threshold value (lower-bound) of precipitation
for active growing season.}

\item{woody}{RasterLayer. A per-pixel percentage of woody vegetation cover.}

\item{c3.crop}{RasterLayer. A per-pixel percentage of cultivated C3 herbaceous vegetation.}

\item{c4.crop}{RasterLayer. A per-pixel percentage of cultivated c4 herbaceous vegetation.}

\item{veg.scale}{Numeric. Scale factor for vegetation cover. Default value = 100.}

\item{veg.index}{Raster* object. Each layer corresponds to a different
temporal window and should match the temporal window and resolution of the climate stacks.}

\item{filename}{Character. Optional output root filename passed to
\code{wric4.teRaster}, default output file type is GeoTiff. If not specified,
output is written to a temporary file.}

\item{...}{Other arguments passed to \code{writeRaster}.}
}
\value{
RasterBrick. Layers correspond to the generated (Layer 1) isoscape and (Layer 2) standard deviation layer.
}
\description{
This is a wrapper function that can calculate an isoscape based on a simple mixing model to percent vegetation cover
  based on user input data and parameters. At minimum a user must provide a temperature and precipitation data.
  The temporal grain and window as well as spatial grain and extent of all input data (climate, vegetation cover,
  vegetation index) must match one another. These inputs would result in a natural herbaceous only isoscape. Users
  can provide woody vegetation data, cultivated vegetation data (crops) and custom crossover thresholds and isotopic endmembers.
  If the users chooses to include crop data, the user must provide and specify a percent c3 crop layer and a percent c4 crop layer.
  Default values for crossover thresholds and isotopic endmembers are listed below.
}
\examples{
# Climate data for Colorado to use in this example
data(COMeanPre)  # mean monthly precipitation (mm)
data(COMeanTmp)  # mean monthly temperature (deg. C)
\donttest{
# Estimate C3/C4 herbaceous layers in 100\\\% grassy world using default thresholds.
herbaceous <- isoscape(COMeanTemp, COMeanPre))}

}
\seealso{
\link[grassmapr]{calcDel13C}, \link[raster]{reclassify}.
}
