% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_grass_pathway.R
\name{map_grass_pathway}
\alias{map_grass_pathway}
\title{Map C4 and C3 proportion of grasses}
\usage{
map_grass_pathway(temp.stack, precip.stack, C4.threshold, GS.threshold,
  precip.threshold, filename = "", ...)
}
\arguments{
\item{temp.stack}{Raster* object. Each layer corresponds to a
different temporal window (e.g., month) for a temperature climate variable.
Object may be single or multi-layer.}

\item{precip.stack}{Raster* object. Each layer corresponds to a different
temporal window (e.g., month) for a precipitation climate variable. Object
may be single or multi-layer.}

\item{C4.threshold}{Numeric. Threshold value (lower-bound inclusive) for
C4 climate temperature.}

\item{GS.threshold}{Numeric. Threshold value (lower-bound inclusive) for
Growing Season temperature.}

\item{precip.threshold}{Numeric. Threshold value (lower-bound inclusive) for
Growing Season precipitation.}

\item{filename}{Character. Optional output root filename passed to
\code{writeRaster}, default output file type is GeoTiff. If not specified,
output is written to a temporary file.}

\item{...}{Other arguments passed to \code{writeRaster}.}
}
\value{
RasterBrick. Layer[[1]] corresponds to percent of grasses with C4
  photosynthetic pathway, Layer[[2]] corresponods to percent of grasses with
  C3 photosynthetic pathway.
}
\description{
Wrapper function to map proportion of grasses with C4 and C3
  photosynthetic pathway.
}
\examples{
\donttest{# Load data for example
data(temp_NA)  # mean monthly temperature (deg. C)
data(prec_NA)  # mean monthly precipitation (mm)

# Specify climate thresholds
C4_temp <- 22    # mean monthly temperature >= 22 deg. C
GS_temp <- 5     # mean monthly temperature >= 5 deg. C
GS_prec <- 25    # mean monthly precipitation >= 25 mm

# Predict percent of grasses by photosynthetic pathway.
grass_map <- map_grass_pathway(temp.stack = temp_NA,
                               precip.stack = prec_NA,
                               C4.threshold = C4_temp,
                               GS.threshold = GS_temp,
                               precip.threshold = GS_prec)

# Plot C4 and C3 grass layers.
plot(grass_map)
}
}
\seealso{
\link[grassmapr]{mask_climate}, \link[grassmapr]{calc_C4_ratio},
\link[grassmapr]{calc_pft_cover}.
}
