% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskC4Climate.R
\name{maskC4Climate}
\alias{maskC4Climate}
\title{Mask C4 climate Based on Temperature and Precipitation masks}
\usage{
maskC4Climate(temp.mask, precip.mask, filename = "", ...)
}
\arguments{
\item{temp.mask}{Raster* object. Mask of grid cells that meet C4 temperature threshold. Object may be single or multi-layer.}

\item{precip.mask}{Raster* object. Mask of grid cells that meet minimum precipitation threshold. Object may be single or multi-layer.}

\item{...}{}

\item{file.name}{Character. Output base filename (highly recommended for large files).}
}
\value{
RasterBrick. One output mask layer indicating C4 climate for each input layer. Output file format is GeoTiff with unsigned integer data type.
}
\description{
Mask C4 climate Based on Temperature and Precipitation masks
}
\examples{

}
