% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskClimate.R
\name{maskClimate}
\alias{maskClimate}
\title{Mask climate Raster values}
\usage{
maskClimate(climate.stack, threshold, filename = "", ...)
}
\arguments{
\item{climate.stack}{Raster* object. Monthly values for a single climate
variable. Object may be single or multi-layer.}

\item{threshold}{Numeric. Threshold value for a single climate variable.}

\item{filename}{Character. Output base filename (required for large files).}

\item{...}{}
}
\value{
RasterBrick. One output mask layer for each input data layer. Output file format is GeoTiff with unsigned integer data type.
}
\description{
This function builds monthly climate masks based on a user-specified
threshold for a single climate parameter (e.g., precipitation, temperature,
etc.).
}
\examples{

}
