% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskClimateVals.R
\name{maskClimateVals}
\alias{maskClimateVals}
\title{Mask Climate Grids Based on Single Parameter}
\usage{
maskClimateVals(climate.stack, threshold, filename = "", ...)
}
\arguments{
\item{climate.stack}{Raster* object. Each layer corresponds to a
different temporal window (e.g., month) for a single climate variable.
Object may be single or multi-layer.}

\item{threshold}{Numeric. Threshold value (lower-bound) for a single climate
variable.}

\item{filename}{Character. Optional output root filename passed to
\code{writeRaster}, default output file type is GeoTiff. If not specified,
output is written to a temporary file.}

\item{...}{Other arguments passed to \code{writeRaster}.}
}
\value{
RasterBrick. One output mask layer for each input data layer. If
  filename is specified, output file format is GeoTiff with unsigned integer
  data type.
}
\description{
Builds climate masks based on a user-specified threshold for a single climate
  parameter (e.g., precipitation, temperature). Grid cells greater than or
  equal to threshold are reclassified as 1, values less than thershold as 0.
}
\examples{
# Climate data for Colorado to use in this example
data(COMeanPre)  # mean monthly precipitation (mm)
data(COMeanTmp)  # mean monthly temperature (deg. C)
\donttest{
# Mask precipitation values >= 25 mm
precip_mask <- maskClimateVals(COMeanPre, 25)}
\donttest{
# Mask temperature values >= 5 deg. C in Colorado
GS_temp_mask <- maskClimateVals(COMeanTmp, 5)}
\donttest{
# Mask temperature values >= 22 deg. C in Colorado
C4_temp_mask <- maskClimateVals(COMeanTmp, 22)}

# Plot monthly growing season temperature masks
plot(GS_temp_mask)

# Plot C4 temperature mask for July only
plot(C4_temp_mask[[7]])
}
\seealso{
\link[grassmapr]{combineMasks}, \link[raster]{reclassify}.
}
