% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_NA-data.R
\docType{data}
\name{precip_NA}
\alias{precip_NA}
\title{Mean Monthly Precipitation Grid for North America}
\format{An object of class \code{"RasterBrick"}, layers correspond to months;
  see \code{\linkS4class{Raster}}.}
\source{
\emph{CRU CL v.2.0}. Climatic Research Unit, University of East
  Anglia: \url{http://www.cru.uea.ac.uk/data/}
}
\usage{
data(precip_NA)
}
\description{
Gridded climatology of 1961-1990 monthly mean precipitation (in mm) at 10'
  spatial resolution. Climatic Research Unit, High-Resolution Gridded
  Dataset.
}
\examples{
# Read in the data
data(precip_NA)
\donttest{
# Plot all data layers
plot(precip_NA)}
\donttest{
# Plot mean monthly precipitation for March
plot(precip_NA[[3]])}

}
\references{
New M, \emph{et al.} 2002. A high-resolution data set of surface
  climate over global land areas. \emph{Climate Research} 21:1-25.
}
\keyword{datasets}
