% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpMeanNA-data.R
\docType{data}
\name{tmpMeanNA}
\alias{tmpMeanNA}
\title{Mean Monthly Temperature Grid for North America}
\format{An object of class \code{"RasterBrick"}, layers correspond to months;
   see \code{\link[raster]{Raster-class}}.}
\source{
\emph{CRU CL v.2.0}. Climatic Research Unit, University of East
  Anglia: \url{http://www.cru.uea.ac.uk/data/}
}
\usage{
data(tmpMeanNA)
}
\description{
Gridded climatology of 1961-1990 monthly mean temperature (in deg. C) at 10'
  spatial resolution. Climatic Research Unit, High-Resolution Gridded
  Dataset.
}
\examples{
# Read in the data
data(tmpMeanNA)
\donttest{
# Plot all data layers
plot(tmpMeanNA)}
\donttest{
# Plot layer 3 only
plot(tmpMeanNA[[3]])}

}
\references{
New M, \emph{et al.} 2002. A high-resolution data set of surface
#'   climate over global land areas. \emph{Climate Research} 21:1-25.
}
\keyword{datasets}
