% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woodyNA-data.R
\docType{data}
\name{woodyNA}
\alias{woodyNA}
\title{Percent woody cover for North America}
\format{An object of class \code{"RasterLayer"}; see
  \code{\link[raster]{Raster-class}}.}
\source{
\emph{MODIS Vegetation Continuous Fields}. Global Land Cover Facility,
  University of Maryland: \url{http://glcf.umd.edu/data/vcf/}
}
\usage{
data(woodyNA)
}
\description{
Gridded percent woody cover layer for North America in the year 2001 at 
  10' spatial resolution. Derived from MODIS Vegetation Continuous Fields 
  tree coverlayer (v006).
  following rules described in Powell \emph{et al.} 2012.
}
\examples{
# Read in the data
data(woodyNA)
\donttest{
# Plot data layer
plot(woodyNA)}

}
\references{
Friedl MA, \emph{et al.} 2010. MODIS Collection 5 global land
  cover: Algorighm refinements and characterization of new datasets.
  \emph{Remote Sensing of Environment} 114:168-182.

Hansen MC, \emph{et al.} 2003. Global percent tree cover at a spatial
  resolution of 500 meters: first results of the MODIS vegetation continuous
  fields algorithm. \emph{Earth Interactions} 7:10.

Powell RL, Yoo EH, and Still CJ. 2012. Vegetation and soil carbon-13
  isoscapes for South America: integrating remote sensing and ecosystem
  isotope measurements. \emph{Ecosphere} 3:109.
}
\keyword{datasets}
