% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_C4_ratio.R
\name{calc_C4_ratio}
\alias{calc_C4_ratio}
\title{Predict C4 Ratio for Grasses}
\usage{
calc_C4_ratio(C4.mask, GS.mask, veg.index = NULL, filename = "", ...)
}
\arguments{
\item{C4.mask}{Raster* object. Each layer represents C4 climate mask for
given temporal window, may be result of  \code{intersect_masks}. Must have
same number of layers as \code{GS.mask}.}

\item{GS.mask}{Raster* object. Each layer represents growing season climate
mask for given temporal window, may be result of \code{intersect_masks}.
Must have same number of layers as \code{C4.mask}.}

\item{veg.index}{Raster* object. Optional layers to weight C4 grass
proportion based on vegetation "greenness" metric (e.g., a vegetation
index). Must match temporal window and number of layers as \code{C4.mask}
and \code{GS.mask}.}

\item{filename}{Character. Optional output root filename passed to
\code{writeRaster}, default output file type is GeoTiff. If not specified,
output is written to a temporary file.}

\item{...}{Other arguments passed to \code{writeRaster}.}
}
\value{
Raster layer. Proportion of grasses with C4 photosynthetic pathway.
}
\description{
Calculates the proportion of grasses with C4 photosynthetic pathway based on
  climate data (and, optionally, vegetation "greenness" weights). Output
  layer represents the C4 proportion of grasses in each grid cell, based on
  ratio of C4-favorable months to growing-season months.
}
\examples{
\donttest{# Calculate C4 ratio for North America, based on climate data only

# C4 and Growing Season (GS) climate masks are results of mask_climate()
C4_ratio <- calc_C4_ratio(C4_mask, GS_mask)

# Calculate C4 ratio, based on climate and monthly vegetation greenness

# Load monthly vegetation index for North America
data(ndvi_NA)
C4_ratio_vi <- calc_C4_ratio(C4_mask, GS_mask, veg.index = ndvi_NA)

# Plot climate-based C4 vegetation ratio
plot(C4_ratio)
}
}
\seealso{
\link[grassmapr]{intersect_masks}, \link[grassmapr]{count_months}, \link[raster]{overlay}.
}
