% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcC4Ratio.R
\name{calcC4Ratio}
\alias{calcC4Ratio}
\title{Calculate C4 Ratio for Grasses}
\usage{
calcC4Ratio(C4.mask, GS.mask, veg.index = NULL, filename = "", ...)
}
\arguments{
\item{C4.mask}{Raster* object of C4 climate masks.}

\item{GS.mask}{Raster* object of Growing Season climate masks.}

\item{veg.index}{Raster* object of vegetation index ("greenness") values, corresponding to same temporal units as C4.mask and GS.mask}

\item{filename}{Output root filename.}

\item{...}{}
}
\description{
Calculate C4 Ratio for Grasses
}
\examples{

}
