% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDel13C.R
\name{calcDel13C}
\alias{calcDel13C}
\title{Calculate d13C isoscape}
\usage{
calcDel13C(pft.cover, d13C.embs, scale = 100, filename = "", ...)
}
\arguments{
\item{pft.cover}{RasterBrick. Percent cover for each plant functional type,
may be result of \code{calcPFTCover}.}

\item{d13C.embs}{Numeric. Vector of length equal to number of layers in
\code{pft.cover}. Values correspond to d13C endmembers for each
\code{pft.cover} vegetation type.}

\item{scale}{Integer. Scale factor for vegetation cover values. Default
value = 100.}

\item{filename}{Character. Optional output root filename passed to
\code{writeRaster}, default output file type is GeoTiff. If not specified,
output is written to a temporary file.}

\item{...}{Other arguments passed to \code{writeRaster}.}
}
\value{
RasterLayer. Values correspond to mean vegetation d13C (per mil) for
  each grid cell.
}
\description{
Applies simple mixing model to percent vegetation cover layers generated by
  \code{calcPFTCover} or \code{calcC4Ratio}. User provides d13C endmember
  values for each input plant functional type. Function may also be applied
  to another endmember statistic, e.g., input endmember standard deviation
  to generate an "uncertainty" layer.
}
\examples{
\donttest{# User provides vector of d13C endmembers corresponding to input vegetation layers
d13C_emb <- c(-12.5, -26.7, -27.0)  # C4 herb, C3 herb, Woody

# Generate d13C isoscape
d13C_iso <- calcDel13C(pft_cover, d13C_emb)

# Plot d13C isoscape
plot(d13C_iso)}

\donttest{# Calculate weighted standard deviation of mean d13C value
d13C_emb_std <- c(1.1, 2.3, 1.7)  # endmember std.dev.
d13C_iso_std <- calcDel13C(pft_cover, d13C_std)

# Plot standard deviation layer
plot(d13C_iso_std)}
}
\seealso{
\link[grassmapr]{calcPFTCover}, \link[grassmapr]{calcC4Ratio},
  \link[raster]{overlay}.
}
