% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPFTCover.R
\name{calcPFTCover}
\alias{calcPFTCover}
\title{Calculate Plant Functional Type Cover Layers}
\usage{
calcPFTCover(C4.ratio, GS.mask, veg.layers = NULL, C4.flag = NULL,
  herb.flag = NULL, scale = 100, filename = "", ...)
}
\arguments{
\item{C4.ratio}{RasterLayer. C4 ratio of grass cover, result of
\code{calcC4Ratio}.}

\item{GS.mask}{RasterBrick. Monthly mask of grid cells that meet growing
season criteria, result of \code{maskClimateVals} and \code{combineMasks}.}

\item{veg.layers}{Raster* object. Each layer corresponds to non-herbaceous
cover or agriculatural crop layers.}

\item{C4.flag}{Numeric. Vector with length equal to number of layers in
\code{veg.layers} object, with C4 vegetation flag = 1,  C3 vegetation flag
= 0.}

\item{herb.flag}{Numeric. Vector with length equal to number of layers in
\code{veg.layers} object, with herbaceous vegetation flag = 1, woody
vegetation flag = 0.}

\item{scale}{Numeric. Scale factor for vegetation cover. Default value = 100.}

\item{filename}{Character. Optional output root filename passed to
\code{writeRaster}, default output file type is GeoTiff. If not specified,
output is written to a temporary file.}

\item{...}{Other arguments passed to \code{writeRaster}.}
}
\value{
RasterBrick. Layers correspond to percent vegetation cover, stacked
  in the following order: Layer 1 = C4 herbaceous, Layer 2 = C3 herbaceous,
  Layer 3, Layer 4, ... = additional vegetation layers provided by user.
}
\description{
Generates plant functional type (PFT) cover layers, starting with an inital
  assumption of 100\% "grassy world". If other vegetation cover layers are
  provided by the user, total grass cover is adjusted by (1) subtracting the
  sum of the other vegetation layers from 100\% grass cover, and (2)
  partitioning the remaining percent grass cover into C3 and C4 components by
  multiplying by the C4 grass ratio (result of \code{calcC4Ratio}). If C3/C4
  crop cover layers are provided, each is added to the respective C3/C4 grass
  cover layers.
}
\examples{
# Load additional vegetation cover layers to use in this example
data(COWoody)  # percent woody vegetation cover in Colorado
data(COC3Crop) # percent C3 crop cover in Colorado
data(COC4Crop) # percent C4 crop cover in Colorado
\donttest{
# Downsample C4_ratio layer to match spatial resolution of other layers
C4_ratio_rs <- resample(x = C4_ratio, y = COWoody, method = 'ngb')

# Create raster stack of other (non-grassy) vegetation layers
veg_layers <- stack(COC4Crop, COC3Crop, COWoody)

# Indicate layers that correspond to C4 vegetation
C4_flag <- c(1, 0, 0)
# Indicate layers that correspond to herbaceous vegetation
herb_flag <- c(1, 1, 0)

# Generate PFT vegetation cover brick
pft_cover <- calcPFTCover(C4_ratio_rs, veg_layers, C4_flag, herb_flag)

# Plot PFT vegetation cover layers
plot(pft_cover)
}
}
\seealso{
\link[grassmapr]{calcC4Ratio}, \link[raster]{overlay}.
}
