% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineMasks.R
\name{combineMasks}
\alias{combineMasks}
\title{Generate Climate Masks Based on Two Parameters}
\usage{
combineMasks(temp.mask, precip.mask, filename = "", ...)
}
\arguments{
\item{temp.mask}{Raster* object. Mask of grid cells that meet minimum
temperature threshold, may be result of \code{maskClimateVals}. Object may
be single or multi-layer.}

\item{precip.mask}{Raster* object. Mask of grid cells that meet minimum
precipitation threshold, may be result of \code{maskClimateVals}. Object
may be single or multi-layer.}

\item{filename}{Character. Optional output root filename passed to
\code{writeRaster}, default output file type is GeoTiff. If not specified,
output is written to a temporary file.}

\item{...}{Other arguments passed to \code{writeRaster}.}
}
\value{
RasterBrick. One output mask layer for each input layer. If filename
  is specified, output file format is GeoTiff with unsigned integer data
  type.
}
\description{
Intersects temperature and precipitation climate masks into a single mask
  layer, for example identifying grid cells that correspond to growing-season
  climate criteria or C4 climate criteria. Grid cells reclassified as 1
  satisify both temperature and precipitation criteria, grid cells classified
  as 0 fail to satisify criteria for at least one variable.
}
\examples{
\donttest{# Generate masks of grid cells satisfying monthly C4-climate criteria
C4_mask <- combineMasks(C4_temp_mask, precip_mask)}
\donttest{
# Generate masks of grid cells that satisfying monthly growing-season climate criteria
GS_mask <- combineMasks(GS_temp_mask, precip_mask)

#Plot monthly growing-season masks
plot(GS_mask)}
}
\seealso{
\link[grassmapr]{maskClimateVals}, \link[raster]{overlay}.
}
