% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMonths.R
\name{countMonths}
\alias{countMonths}
\title{Count Number of Months in Climate Mask Layers}
\usage{
countMonths(climate.mask, filename = "", ...)
}
\arguments{
\item{climate.mask}{Raster* object. Mask of grid cells that meet climate requirements. Object should be  multi-layer (i.e., multi-date).}

\item{filename}{Character. Output base filename (required for large files).}

\item{...}{}
}
\value{
RasterLayer. Pixel values correspond to number of months that climate requirements are met.
}
\description{
Count Number of Months in Climate Mask Layers
}
\examples{

}
