% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_climate.R
\name{mask_climate}
\alias{mask_climate}
\title{Generate Climate Masks Based on Two Parameters}
\usage{
mask_climate(temp.stack, temp.threshold, precip.stack, precip.threshold,
  filename = "", ...)
}
\arguments{
\item{temp.stack}{Raster* object. Each layer corresponds to a
different temporal window (e.g., month) for a temperature climate variable.
Object may be single or multi-layer.}

\item{temp.threshold}{Numeric. Threshold value (lower-bound, inclusive) for
temperature variable.}

\item{precip.stack}{Raster* object. Each layer corresponds to a
different temporal window (e.g., month) for a precipitation climate variable.
Object may be single or multi-layer.}

\item{precip.threshold}{Numeric. Threshold value (lower-bound, inclusive) for
precipitation variable.}

\item{filename}{Character. Optional output root filename passed to
\code{writeRaster}, default output file type is GeoTiff. If not specified,
output is written to a temporary file.}

\item{...}{Other arguments passed to \code{writeRaster}.}
}
\value{
RasterBrick. One output mask layer for each input layer. If filename
  is specified, output file format is GeoTiff with unsigned integer data
  type.
}
\description{
Generates climate masks based on two climate variables, typically
  temperature and precipitation. Grid cells reclassified as 1 satisfy both
  temperature and precipitation criteria; grid cells classified as 0 fail
  to satisfy criteria for at least one variable.
}
\examples{
\donttest{# Load data for example
data(temp_NA)
data(precip_NA)

# Specify Growing Season (GS) thresholds
gs_temp <- 5   # mean monthly temperature >= 5 deg. C
gs_prec <- 25  # mean monthly precipitation >= 25 mm

# Generate monthly masks of grid cells that satisfy GS-climate criteria
GS_masks <- mask_climate(temp_NA, gs_temp, precip_NA, gs_precip)

#Plot monthly growing-season masks
plot(GS_masks)
}
}
\seealso{
\link[grassmapr]{mask_grids}, \link[grassmapr]{intersect_masks},
\link[raster]{overlay}.
}
